// src/app/heroes/interfaces/heroes.interface.ts
// Generated by https://quicktype.io

export interface Heroe {
  id:               string;
  superhero:        string;
  publisher?:        Publisher;
  alter_ego:        string;
  first_appearance: string;
  characters:       string[];
  description:      Description[];
  alt_img?:         string
}

export interface Description {
  id:   number;
  text: string;
}

export enum Publisher {
  DCComics = "DC Comics",
  MarvelComics = "Marvel Comics",
}